REF: misplaced astype tests
